Move impteam reset tests to separate file, support PUKtoNoPUK case
